+++ src/quadtree/quadtree.rs
@@ -3,7 +3,7 @@ use crate::quadtree::aabb::AABB;
 use crate::quadtree::error::InsertError;
 use crate::quadtree::free_list;
 use crate::quadtree::free_list::{FreeList, IndexType};
-use crate::quadtree::node::{Node, NodeElementCountType};
+use crate::quadtree::node::Node;
 use crate::quadtree::node_data::{NodeData, NodeIndexType};
 use crate::quadtree::node_info::NodeInfo;
 use crate::quadtree::node_list::NodeList;
@@ -14,9 +14,6 @@ use smallvec::SmallVec;
 use std::collections::HashSet;
 
 // TODO: Add range query: Query using intersect_aabb() or intersect_generic()
-
-/// Each node must have less than the maximum allowed number of elements.
-const MAX_NUM_ELEMENTS: NodeElementCountType = 16; // TODO: Make parameter of tree
 
 /// We use this value to determine whether a node can be split.
 const SMALLEST_CELL_SIZE: i32 = 1; // TODO: Make parameter of tree
@@ -66,6 +63,8 @@ where
     free_node: free_list::IndexType,
     /// Stores the maximum depth allowed for the quadtree.
     max_depth: u32,
+    /// Stores the maximum number of elements allowed before a node splits.
+    max_num_elements: u32,
 }
 
 impl<ElementId> QuadTree<ElementId>
@@ -73,10 +72,10 @@ where
     ElementId: ElementIdType,
 {
     pub fn default() -> Self {
-        Self::new(QuadRect::default(), 8)
+        Self::new(QuadRect::default(), 8, 16)
     }
 
-    pub fn new(root_rect: QuadRect, max_depth: u32) -> Self {
+    pub fn new(root_rect: QuadRect, max_depth: u32, max_num_elements: u32) -> Self {
         Self {
             element_ids: FreeList::default(),
             element_rects: FreeList::default(),
@@ -85,6 +84,7 @@ where
             root_rect,
             free_node: free_list::SENTINEL,
             max_depth,
+            max_num_elements,
         }
     }
 
@@ -118,7 +118,7 @@ where
                 };
 
                 let can_split = leaf.can_split_further(SMALLEST_CELL_SIZE, self.max_depth);
-                let node_is_full = element_count >= MAX_NUM_ELEMENTS;
+                let node_is_full = element_count >= self.max_num_elements;
 
                 let must_store_element = !node_is_full || !can_split;
                 if must_store_element {
@@ -589,7 +589,7 @@ where
     where
         T: IntersectsWith<AABB>,
     {
-        let capacity = leaves.len() * MAX_NUM_ELEMENTS as usize;
+        let capacity = leaves.len() * self.max_num_elements as usize;
         let mut node_set = HashSet::with_capacity(capacity);
 
         while !leaves.is_empty() {
@@ -627,7 +627,7 @@ where
 #[cfg(test)]
 pub(crate) fn build_test_tree() -> QuadTree {
     let quad_rect = QuadRect::new(-20, -20, 40, 40);
-    let mut tree = QuadTree::new(quad_rect, 1);
+    let mut tree = QuadTree::new(quad_rect, 1, 1);
     // top-left
     tree.insert(QuadTreeElement::new(1000, AABB::new(-15, -15, -5, -5)))
         .expect("insert should work");
@@ -672,7 +672,7 @@ mod test {
     #[test]
     fn cleanup_works() {
         let quad_rect = QuadRect::new(-20, -20, 40, 40);
-        let mut tree = QuadTree::new(quad_rect, 1);
+        let mut tree = QuadTree::new(quad_rect, 1, 1);
         // top-left
         tree.insert(QuadTreeElement::new(1000, AABB::new(-15, -15, -5, -5)))
             .expect("insert should work");